requirements.txt
apache-beam==2.27.0
avro-python3==1.9.2.1
certifi==2020.12.5
chardet==4.0.0
crcmod==1.7
dill==0.3.1.1
docopt==0.6.2
fastavro==1.3.1
future==0.18.2
grpcio==1.35.0
hdfs==2.5.8
httplib2==0.17.4
idna==2.10
jsonlines==2.0.0
mock==2.0.0
numpy==1.20.0
oauth2client==4.1.3
pbr==5.5.1
protobuf==3.14.0
pyarrow==2.0.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pydot==1.4.1
pymongo==3.11.3
pyparsing==2.4.7
python-dateutil==2.8.1
pytz==2021.1
requests==2.25.1
rsa==4.7
six==1.15.0
typing-extensions==3.7.4.3
urllib3==1.26.3
